Move create page state into CreateAssetProvider
