+++ components/ChooseCollection/index.tsx
@@ -1,4 +1,4 @@
-import { useState, useEffect, Dispatch, SetStateAction } from 'react';
+import { useState, useEffect, FC, Dispatch, SetStateAction } from 'react';
 import {
   Title,
   SubTitle,
@@ -6,36 +6,27 @@ import {
   ElementTitle,
   ErrorMessage,
 } from '../CreatePageLayout/CreatePageLayout.styled';
-import CollectionsCarousel, {
-  NewCollection,
-  CarouselCollection,
-} from '../CollectionsCarousel';
+import CollectionsCarousel from '../CollectionsCarousel';
+import Button from '../Button';
+import {
+  useAuthContext,
+  useCreateAssetContext,
+  CREATE_PAGE_STATES,
+} from '../../components/Provider';
 import { getAuthorsCollections, Collection } from '../../services/collections';
-import { useAuthContext } from '../../components/Provider';
-import Button from '../../components/Button';
 
-type Props = {
-  collectionsList: Collection[];
-  setSelectedCollection: Dispatch<SetStateAction<CarouselCollection>>;
-  setNewCollection: Dispatch<SetStateAction<NewCollection>>;
-  setIsUncreatedCollectionSelected: Dispatch<SetStateAction<boolean>>;
-  selectedCollection: CarouselCollection;
-  newCollection?: NewCollection;
-  goToCreateTemplate: () => void;
-  setCollectionsList: Dispatch<SetStateAction<Collection[]>>;
-};
-
-const ChooseCollection = ({
-  collectionsList,
-  setSelectedCollection,
-  setNewCollection,
-  setIsUncreatedCollectionSelected,
-  selectedCollection,
-  newCollection,
-  goToCreateTemplate,
-  setCollectionsList,
-}: Props): JSX.Element => {
+const ChooseCollection: FC<{
+  setPageState: Dispatch<SetStateAction<string>>;
+}> = ({ setPageState }) => {
   const { currentUser, isLoadingUser } = useAuthContext();
+  const {
+    setSelectedCollection,
+    setNewCollection,
+    setIsUncreatedCollectionSelected,
+    selectedCollection,
+    newCollection,
+  } = useCreateAssetContext();
+  const [collectionsList, setCollectionsList] = useState<Collection[]>([]);
   const [fetchError, setFetchError] = useState<string>('');
   const [noneSelectedError, setNoneSelectedError] = useState<string>('');
   const [isLoadingCollections, setIsLoadingCollections] = useState<boolean>(
@@ -70,7 +61,7 @@ const ChooseCollection = ({
       );
     } else {
       setNoneSelectedError('');
-      goToCreateTemplate();
+      setPageState(CREATE_PAGE_STATES.CREATE_TEMPLATE);
     }
   };
